0.1.1/requirements.txt
streamlit>=1.29.0
pandas>=2.0.0
plotly>=5.0.0
pyyaml>=6.0
requests>=2.25.0